Fix some errors, add MathJax, add a draft post.
The errors fixed include the disqus fragment that would include
disques even outside a production environment and a fix to the
css for a <sup> tag (it wasn't being show as superscript).

Added MathJax support which necessitated overriding the theme's default
layout. Added a draft post that uses the MathJax support.